Added a proper progress calculation and it shows the progress in the console somewhat smooth and bugless (not always)

+++ MultithreadDownloader/DownloadThread.cs
@@ -13,7 +13,7 @@ namespace MultithreadDownloader
         public string ThreadName;
         public string Status;
         public long ProgressAbsolute;
-        public int ProgressRelative;
+        public float ProgressRelative;
         public DownloadThread(long start, long end, string threadname)
         {
             Start = start;
